Fix bug #452: loadPixels & co. for retina displays
First, this patch makes _pixelDensity -> pixelDensity so that
it is no longer private and can be used appropriately when handling
images.

pixelDensity is needed for proper iteration of pixels[]. Anyone
wishing to do fast get/set of pixels[] will need to account for
pixelDensity when determining the total number of pixels available
and the location of said pixels.

get() and set() now work like an end user would expect; they both
respect the size of the image/sketch as defined by the user. set()
will fill an appropriate number of pixels depending on the pixel
density of the window (for example, 4 pixels on a retina display
when a single pixel is requested) so that an end user doesn't have
to worry about the pixelDensity at all. same goes for get.

To properly scale images down from retina to regular size, get()
now uses canvas's drawImage to handle scaling.

Examples & documentation are updated to handle this the new
name for pixelDensity and to iterate over pixels[] properly so
that all examples look the same whether viewed on a retina or
normal display.

Additionally, P5.Image gets pixelDensity = 1 by default so that
it can be used nicely with image/pixels.js.