shader refactor wip
begins to build out a more robust p5.Shader class in shader.js.

the motivation for this refactor is that in 2017 it seems that
shaders are the core of any gl implementation. so it makes sense
to have a lot of the heavy lifting there. this patterns adopts
some ideas from Processing's shader class, with some adaptations
for our codebase (for example, that we're not using VBOs and
that we have two distinct render modes in GL).

the material cache now maps strings (as before, names of the shaders
by connecting the vertex and fragment shader with a pipe) to the
p5.Shader objects. retain mode and immediate mode both make use
of this new abstraction. when possible, methods are called
directly on the shader object to set uniforms and attributes.
the "bind" method now prepares common shader attributes for
rendering: the projection and modelview matrices, etc.

binding buffers is still handled by the appropriate render mode.
in the future, I think there's room for the shader class to have
its own draw method that takes an appropriate set of buffers for
various properties. let's see if that works in a future commit.

future todo items are:
- moving texture handling from texture.js to shader.js
- more general handling of drawing buffers between retain
  and immediate mode. it seems there is more overlap here
  than I inititally thought.
- revise naming of _getCurShaderId and establish better
  pattern and API for setting the current shader in the
  various places it might need to be accessed. (for example,
  p5Shader.useProgram seems a little bit odd on its own
  but is used in that way in a few cases--ideally this should
  only be set via renderer, when it picks a p5.Shader object,
  maybe -- not sure about this one yet)