optimizations for p5.Image + textures, use map for textures in renderer
adds 'modified' field to p5.Image for tracking is updates to textures
are needed. it's not fool proof: if someone draws to the underlying
canvas in the implementation, it won't get marked for update. but any
interactions with the image pixels will flag it as having been modified.
API user can then set modified to false when they've examined the value.

change array of textures to an object/map in p5.Renderer.GL per David's
suggestion.