fix ambient light, gl error when lighting is set but no texture
ambientLight was not incrementing the # of lights in the scene as it
should have been. must have been accidentally deleted during code cleanup
from a prior patch, oops!

and: importantly, restores the empty texture work Stalgia had added but
using the new p5.Texture architecture. a default texture is loaded
and used (pure white, 1x1 pixels from an in-memory p5.Image) when the
user has not supplied a texture to the shader.